backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-docx
openai
python-dotenv
